Create PickUpBottle.py
This program spins the base 45 degrees,  moves the arm down 90 degrees, closes the gripper to pick up a bottle and goes back to the starting position releasing the bottle again. It then repeats this cycle until stopped. This was a prove of concept for using this robot in a manufacturing environment to pick up parts from a conveyor belt and move them to another place.